Fix warnings

See merge request jschoeberl/ngsolve!732

+++ linalg/blockjacobi.hpp
@@ -171,7 +171,7 @@ namespace ngla
     Array<MemoryUsage> GetMemoryUsage () const override
     {
       int nels = 0;
-      for (int i = 0; i < blocktable->Size(); i++)
+      for (size_t i = 0; i < blocktable->Size(); i++)
 	{
 	  int bs = (*blocktable)[i].Size();
 	  nels += bs*bs;
